Change data type for longitude and latitude to float

+++ smartlist/db/migrate/20160504152955_create_favorite_stores.rb
@@ -3,8 +3,8 @@ class CreateFavoriteStores < ActiveRecord::Migration
     create_table :favorite_stores do |t|
       t.text :name
       t.string :address
-      t.decimal :latitude
-      t.decimal :longitude
+      t.float :latitude
+      t.float :longitude
 
       t.timestamps null: false
     end
